autorizando novas rotas

+++ app/Http/Controllers/Inscricao/CupomDeDescontoController.php
@@ -36,9 +36,9 @@ class CupomDeDescontoController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {
+    {   
         $evento = Evento::find($request->evento_id);
-
+        $this->authorize('isCoordenadorOrComissaoOrganizadora', $evento);
         $validadeData = $request->validate([
             'identificador' => 'required',
             'quantidade'    => 'required',
@@ -111,6 +111,7 @@ class CupomDeDescontoController extends Controller
     {
         $cupom = CupomDeDesconto::find($id);
 
+        $this->authorize('isCoordenadorOrComissaoOrganizadora', $cupom->evento);
         // Checar se o cupom foi aplicado em alguma inscrição antes de excluir
         $cupom->delete();
